requirements.txt
fastapi
uvicorn[standard]
rembg
pillow
python-multipart
onnxruntime
numpy
opencv-python
pydantic-settings
python-dotenv